move TimelineEntryModel.safeKey to domain data
